pyproject.toml
[project]
name = "agdsw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langgraph>=1.0.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "streamlit>=1.39.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "sqlalchemy>=2.0.45",
]
